enviroment/backend_requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.8
attrs==25.1.0
audio-recorder-streamlit==0.0.10
autogen==0.7.1
av==13.1.0
backoff==2.2.1
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
CacheControl==0.14.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
durationpy==0.9
email_validator==2.2.0
elastic-transport==8.17.1
elasticsearch==8.17.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fast-depends==2.4.12
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.17.0
firebase-admin==6.6.0
FLAML==2.3.3
flatbuffers==25.1.24
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gensim==4.3.3
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core>=2.11.0,<3.0.0
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore>=2.19.0,<3.0.0
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos>=1.59.0,<2.0.0
greenlet==3.1.1
grpcio>=1.48.1,<2.0.0
grpcio-status>=1.48.1,<2.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.33.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jiter==0.8.2
Jinja2==3.1.6
joblib==1.4.2
jpype1==1.5.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
konlpy==0.6.0
kubernetes==32.0.0
langchain==0.3.17
langchain-chroma==0.2.1
langchain-community==0.3.16
langchain-core==0.3.41
langchain-openai==0.3.3
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langcodes==3.5.0
langgraph==0.3.5
langgraph-checkpoint==2.0.17
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.12
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marisa-trie==1.2.1
marshmallow==3.25.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.10
mistune==3.1.2
mltk==0.0.5
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.61.0
openai-whisper==20240930
opencv-python==4.11.0.86
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
ormsgpack==1.9.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
pip==25.0.1
platformdirs==4.3.6
posthog==3.11.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf>=3.19.5,<5.27
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyautogen==0.7.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.11.0
Pympler==1.1
PyMuPDF==1.25.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytube==15.0.0
pytubefix==8.12.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==26.2.1
ragas==0.2.14
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.29.0
sentence-transformers
Send2Trash==1.8.3
setuptools==75.8.2
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.0
streamlit==1.43.2
streamlit-option-menu==0.4.0
streamlit-webrtc==0.53.0
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets>=14,<15
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp